requirements.txt
click==5.1
docopt==0.4.0
httplib2==0.9.2
isodate==0.5.4
mandrill==1.0.57
ordereddict==1.1
python-dateutil==2.4.2
python-dotenv==0.3.0
pytz==2015.7
requests==2.8.1
six==1.10.0
wheel==0.24.0
